Added functionality to (un)freeze the texture plane
Summary:
(1) Added functionality to freeze the texture plane, which is useful when its   
geometry and/or its texture is updated and we don't want to show intermediate   
states to the user.
(2) Added texture shift and texture growth functionality, which is generally    
applicable, but can also be used during an update to correctly display still    
old texture on already new geometry.                                            
(3) A few cosmetic changes.

Test Plan:
(1) Resizing an inl/crl/z-slice (Display->Position) with multiple attributes    
in od_osg while freezing the texture plane (to be committed next), and seeing   
that its display does not change during a freezing stage, where it did before.
(2) Tested by adding them as new options in example_textureplane. Also applied  
in od_osg to display old texture on new geometry, and seeing that (finally)     
the texture appears undeformed at the correct world position.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D1016