Enabled the mouse pickability of PolyLines.
Summary:
Added functionality to service intersection traversals in order to make         
PolyLines pickable. This does not happen automatically, because these           
lines are added to the scene on-the-fly during the cull stage. Also added       
a bounding sphere computation to optimize culling and ascertain that mouse      
picks on a PolyLine are not clipped away.

Test Plan:
Move the mouse pointer over the PolyLine that displays a 2D Horizon and         
see that now the line information is shown in the status bar of OpendTect.

Reviewers: ding

Reviewed By: ding

Differential Revision: https://phabricator.dgbes.com:8443/D1616