Added rotation functionality to TexturePlane class
Summary:
Added TexturePlane class functionality to rotate TexturePlane over arbitrary
angles, where it had to be perpendicular to a coordinate frame axis so far.

Test Plan:
Experimented with new --angle option that was added to example_textureplane.
Also tested in OpendTect which needs it when displaying prestack data in a
3D scene flatviewer being rotated perpendicular to an arbitrary 2D line.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D1931