Customized TabPlaneDragger to override TranslatePlaneDragger event defaults.
Summary:
Middle mouse button for perpendicular-to-plane translation of TabPlaneDragger
is replaced by left mouse button. Also added functionality to specify the
mod key masks for in-plane and perpendicular-to-plane translation.
(REPLACES REVISION D2288 WHICH WAS ACCIDENTALLY OVERWRITTEN BEFORE REVIEW)

Test Plan:
In combination with changes in visdepthtabplanedragger.cc, the inl-crl-z
draggers now show identical manipulation behavior as in Coin3D. This
behaviour is correctly adjusted if the default mod key masks for in-plane
and perpendicular-to-plane translation are modified via the properties
dialog or the advanced settings.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D2343