Adapted all osgGeo texture functionality for viewers with redraw-on-demand policy.
Summary:
No longer allowed to ask for continuous update traversal, because this makes
the redraw-on-demand policy of viewer void. In order not to introduce new bugs,
we keep the "poll-based" updating code the way it is. We only add a mechanism
that requests for redraw if there is any chance that an update is necessary.
The polling during the following update stage will learn whether that actually
is the case. We sytematically located all update "variables" in our code, and
request a redraw when these are raised. A generalized callback functionality
for non-osg callbacks was needed to pass the redraw requests from auxiliary
texture objects to the real scene graph nodes.

Test Plan:
Did various textural user modifications on inlines and 2D lines in OpendTect,
with and without shaders, to see whether it still yields an immediate update
in the 3D scene. Also put breakpoints in OSG to assure that no longer redraws
are done in absence of any user action with only these texture objects in the
scene. Also tested osgGeo/texture_plane_example, which is known to fail if not
continously updating until texture info is available, which is not at startup
right away.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D2816