Added functionality to customize mouse interaction with six TabPlaneDraggers from which the TabBoxDragger is built.
Summary:
(1) TabBoxDragger now build from osgGeo/TabPlaneDraggers, so that it can start
passing access to their mouse interaction customization functions.
(2) Added functionality to figure out from which plane of the TabBoxDragger a
received MotionCommand originates. Also put them in a consistent "front-back"
order with respect to the dimensions of the local coordinate space. This will
be used by OpendTect to (optionally) interpret in-depth translation of the
dragger box as a box resize event.

Test Plan:
(1) Used OpendTect to verify that the original OSG mouse interaction for the
in-depth and in-plane translation of the dragger box can be overruled.
(2) Functionality applied in OpendTect and concluded that the pulling and
pushing of each of the six planes shows consistent behaviour when resizing
the dragger box via in-depth translation.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D2871