Bugfix: No refresh for texture (attribute) change on TexturePlane in OpendTect.
Summary:
No refresh for texture change on TexturePlane in OpendTect. This was introduced
by Phabricator revision D2816 (redraw-on-demand). The bug is caused by a very
tiny optimization in setUpdateVar(.,.), that is invalid in this case and
decided to be removed from all implementations. It could be made valid with
an alternative code change in this case, but then it would remain a potential
source of trouble.

Test Plan: Texture (attribute) change in OpendTect does now refresh again on screen.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D2880