Bugfix: One-marker sets with constant screen size disappeared when zooming in, switching them off, zooming out and switching them on again.
Summary:
One-marker sets with constant screen size disappear when zooming in, switching
them off, zooming out and switching them on again. The zoom-in makes their
world size very small at switch-off time, and next the zoom-out makes them
suffer from small feature culling at switch-on time. There is no chance to
adjust the marker size, because the cull visitor does not come by during the
switch-off and neither afterwards, unless one manages to zoom-in heavily on
the marker again. Many potential solutions have been considered, but deactivate
culling for one-marker sets is the only one feasible. Note that the problem
does not appear for two-and-more-marker sets, because the distance between the
markers is constant in world space and not in screen space.

Test Plan:
Zoom-in on the yellow edit knots of e.g. a faultstick set in OpendTect. Change
to view mode will switch them off, zoom out and change back to act mode again,
and after the fix the yellow edit knots did show up again.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3001