Functionality to minimize texture color vagueness (Mantis 2542)
Summary:
(1) The combination of mipmap filtering, color mapping and a fragment shader
moves the locally averaged texture color towards the middle value of the color
map. Solved by adding extra texture channel(s) with all values squared, from
which the shader can derive the local standard deviation of the texture values
and use that as offset into a texture containing the color scale space(s) of
the color map(s). Note that texture sizes are passed to the shader via uniforms,
because the textureSize(.) GLSL command relies on OpenGL 3.0. Direct mipmap
level queries even require OpenGL 4.3.
(2) Non-isotropic filtering causes vagueness if texels are not projected in
a square-like shape onto screen space, either because the texture sampling is
out-of-balance in horizontal and vertical direction or because the viewing
angle is far from perpendicular. Solved by enabling hardware anisotropic
filtering in OSG, which does the job provided that the graphics card supports
it.
(3) Have LayeredTexture.cpp use same snprintf(.) definition as LayerProcess.cpp
to be consistent and avoid future CDash problems.

Test Plan:
Functionality tested via OpendTect for TexturePlanes and TexturePanelStrips,    
at different zoom distances, viewing angles and depth stretch, for different    
color maps and for attributes with and without undefined areas, with enabled    
and disabled shader. Great improvement in color stability and sharpness.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3355