Functionality to support multi-slice images in layered textures.
Summary:
1) Added functionality to store multi-slice images and select one slice.
A selection change will be fast with shaders, but does require resampling
the composite texture without.
(2) Added functionality to change the order of the data dimensions of a stored
multi-slice image. This enables the image to be sliced in the non-primary
directions as well. The current image stride is not usable for all data orders,
which means that one direction (z-slice in OpendTect) requires copying of the
image tiles of the selected slice. If necessary, performance can be improved
either by multi-threading or maintaining differently ordered duplicates of
each multi-slice image.
(3) Adapted the rescaling of small non-power-of-two textures to (1) and (2).
This also involves copying of images every time when slice and/or data order
are changing. It does not really hurt performance, because it is done only
for small images (to improve tile size restrictions of multi-layer textures).
(4) Replaced obsolete image tile copy functionality for image stride emulation
by an optimized image copy function that supports all image data orders.

Test Plan:
Tested in OpendTect using the new orthogonal slice implementation behind the
tracking cube and plane. Both with and without shaders. Dragging the slice
forward and backward in three different tracker plane directions, while QC-ing
the texture and make sure it does not differ from OpendTect-Coin. The correct
rescaling of small non-power-of-two textures under these circumstances was
tested by temporarily appointing all textures as small non-power-of-two.
Whether the existing layered texture functionality kept working was tested
by continuously monitoring textures on plane sections in OpendTect.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3544