Normal vector pointed in opposite direction if texture axis are swapped.
Summary:
Normal vector pointed in opposite direction if texture axis are swapped.
Bugfix for planar textures in OpendTect, which were insensitive to the diffuse
reflectivity of the display properties dialog if use of shaders is disabled.

Test Plan:
Planar textures in OpendTect are now sensitive to the diffuse reflectivity
of the display properties dialog, also if use of shaders is disabled. osgGeo's
textureplane example does not swap texture axis, but still shows diffuse
reflectivity, also if shaders are disabled. In the TexturePanelStrip class,
the normal vector direction was already depending on whether texture axis 
are swapped.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D3737