Make osgGeo polylinenode work for LINES definition
Summary:
The current polyline node class only work for LINE_STRIP definition in osg::PrimitiveSet. It will generate a wrong result once the primitiveset defined as LINES. This is happended
in the od intersection lines between fault and other objects.
The changes make it work for LINES. for example:
012345 it will generate 3 3d lines 01,34,and 45.
0112344567 it will generate line strip 012, 345 and line 67.

By the way, the changes has also fixed the potential bug in reScale..() function if the geometry has multiple primitive sets

Test Plan: Compiles fine and Tests in od and polylines working fine.

Reviewers: ranojay

Reviewed By: ranojay

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D5582