New interfaces in osgGeo for PDF3D
Summary:
The reason added these new interfaces is that we want to get the geometry information in od viswell.
Since Arnaud wants the repeat log in PDF3D, this revision keeps the repeat interface. We can remove it after discussion.

Test Plan: Compiles fine and Tests locally with od fine

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D6538