od- bug fixing scaler bar hang line length
Summary:
This revision is same as D6859.
The scaler bar hang line is the distance between scaler bar and annotation number. It should not depended on the character size but on the bar's width/height,
because when rendering scaler bar in heigh DPI, the character size is changed and the hang line will be longer than previous DPI.

Test Plan: Compiles and Tests fine

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D6896