Bugfix: Dragger is never released if modkey is pressed at mouse release while it was not at mouse press. (PORTING FROM od TO od5.0 UNCHANGED)
Summary:
Bugfix: Dragger is never released if modkey is pressed at mouse release while
it was not at mouse press. The setInactivationModKeyMask() functionality of
revision r344/D2907 had to be more precise in the events it is blocking.

Test Plan:
Tested in OpendTect. The yellow fault knot now changes back again from arrow
shape to cube shape when we just start pressing e.g. the Ctrl-key during the
dragging of the knot. Also verified that it still requires only a single
Ctrl-click to remove the knot, because that was the primary reason why the
setInactivationModKeyMask() functionality was introduced.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D7135