Bugfix: Clicking in scene results in huge delays if it contains hundreds of draggers.
Summary:
Bugfix: Clicking in scene results in huge delays if it contains hundreds of
draggers. It turns out that every OSG dragger intersects the camera ray with
all scene objects to get an intersection list (osgManipulator/Dragger.cpp:371).
view->computeIntersections(ea ,intersections,_intersectionMask). I don't see
why this list cannot be shared or constructed only if the OSG dragger is under
the mouse pointer itself. My solution overloads the traverse() function to
test the latter condition in advance. We may ask Robert Osfield whether the
current way of implementing this is a necessity or a bug. However, I guess
he did not do the original implementation of the draggers, and the rest of
the community does not use that many draggers in one scene.

Test Plan:
Tested in OpendTect using F3_Demo. I used Fault_A with every knot duplicated
X times (by a code hack) and I used "Example random contour demo 1 ->MFS4"
as it is. Both add hundreds of draggers to the scene. Varying X proves that
the original order of computation is quadratic with the number of draggers.
The bugfix reduces this to a linear order. No huge delays anymore when
clicking in the scene, while clicking on a specific dragger to drag it
does still work.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D7700