Functionality to elevate rotated-to-screen text billboards. (PORTING FROM od TO od5.0 UNCHANGED)
Summary:
Added derived Text class with functionality to elevate rotated-to-screen text
billboards. When looking at a horizon surface from the top, its contour labels
start disappearing behind the surface in rotate-to-screen mode. An angular
elevation with respect to a reference plane can solve this. In perspective
view however, this introduces a position-dependent shear deformation of the
labels in screen space. If the angle between billboard and screen does
not become too big, this can be corrected almost perfectly by a linear
transformation of the original billboard.

Test Plan:
Tested within OpendTect using a not-too-flat horizon with contour labels.
When choosing a proper elevation angle, the labels remain fully visible
without serious distortion for any camera angle that makes sense during
contour viewing. The whole elevation matrix calculation is the result of
prolonged trial and error. It was tested for different elevation angles
and many camera positions, left and right aligned labels, perspective and
orthographic projection, and for surveys with reversed inl-crl sense as
well.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D7777