od-osggeo bug 3873 fixing
Summary:
The bug is caused by delayed Markerset traverse. in od side, after we add a set of markers by position( vertex array), color..., and then call markerset->forcedraw().
In vishorzondisplay, it tries to  call turnMarkerOn, it return false, because in osggeo side the update traverse is not done yet.
from another point of the forcedraw function, I think "forcedraw" is to draw without any condition. So changed a little of this function and fixed the bug 3873.

Test Plan: Compiles and Tests fine

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D7609