od6.0:osggeo - improvement on polygon selection
Summary:
This is the same revision D8492 in od 
Added osggeo polygonselection in which we allow mouse single click and double click work in both of polygonselection and visevent.
The previous version has bug in mouse single click event.

Test Plan: Compiles fine and Tests fins in fault,faultstickset and horizon tracking

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D8493