Bugfix: Current mod key mask index could change during dragging.
Summary:
Current mod key mask index could change during dragging. This may cause a
change of dragging operation during dragging in case mod keys are pressed
too late or released too early.

Test Plan:
Tested with the random track dragger of OpendTect. No longer possible to start
for example a plane drag in-depth and continue it as a rotation as soon as the
Ctrl button is pressed, in which case no rotation axis was drawn either.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D8612