Stopped relying on x and y positions in scroll-event.
Summary:
The x and y are not set on scroll event, so the code relied on the
previous values. When using a trackpad, the x and y were sometimes set to
values that came from some other touch coordinate system and gave random results.

This revision makes the manipulator record the position of every mouse move, and
use the last known position as a basis for the scroll event's zoom.

Test Plan: Tested with both mouse and track-pad, and works fine.

Reviewers: jaap

Reviewed By: jaap

Differential Revision: https://phabricator.dgbes.com:8443/D8924