Enabled reuse of threads
Summary:
The texture generation made new threads every time. As thread creation is expensive, and threads are cheap to keep once they are created, I made
a pool of threads that are shared among all textures.

Test Plan: Tested with inlines on dgb20 to see that it works in a massively parallel environment.

Reviewers: jaap

Reviewed By: jaap

Differential Revision: https://phabricator.dgbes.com:8443/D9024