Bugfix: composite image regenerated twice after geometry change.
Summary:
Bugfix: If shading is off, the necessary composite image is regenerated twice
after an update of the TexturePanelStrip geometry. Composite image generation
is a speed bottleneck. Issue was caused by the assignTextureUnits() function,
which is called twice if TexturePanelStrip geometry changes. It did trigger a
a SetupStateSet update, even if the texture unit assignment stays the same.

Test Plan:
Tested in OpendTect with shading off. Scene contains a textured random line
at highest resolution. The waiting time after moving the random line decreased
and the debugger proves that the composite image is no longer regenerated
twice. The chance that another workflow is accidentally relying on the
SetupStateSet update as done by assignTextureUnits() is considered very
small. If this is going to be reported nevertheless, the _updateSetupStateSet
variable needs to be set to true somewhere else.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D9135