osgGeo: Fixed indexing problem for Polylines with multiple primitivesets
Summary:
Basically had to revert the changes done in the last revision. What was fundamentally going wrong was that pidx is the index of the coordinates in the prinitive set, so pidx+1 does not necessarily mean that it is the next coordinate index e.ge

ps{ 30, 40, 50, 60 };

So pidx0 = 30. but pidix1 should not be pidx+1 i.e 31, rather it should be 40.

so the order of pidx is determined by the stripidx.

Howver the initial problem of handling null vectors is that, the last point p1,and p2, are same, so instead of using the last vector i.e p2 - p1, I would simply use the immediate lats one by reversing it.

Test Plan: seems to work with all kinds of intersections

Reviewers: ding

Reviewed By: ding

Subscribers: raman

Differential Revision: https://phabricator.dgbes.com:8443/D9150