osggeo branch-Added two new functions
Summary:
This is the same revision as D9442 in trunk 
There is bug existing when remove a marker from a markerset, the on/off array is not updated which will lead to wrong on/off of markers in displaying.
The function: removeMarkerOnOff() fixes this problem although it is not the best solution. The best solution I think we need a function called removeMarker() in which we remove the element
from osg array. but we can't,because the osggeo array is linked to od coordinates Typeset<coord3>. in currnt design, we do remove element from od array in vismarkerset.

Test Plan: Compiles fine. and Tests also fine with os pickset removing and adding in the option of displaying at section.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D9445