[OD osggeo branch] Fixing bug found in Intel Parallel Inspector
Summary:
The revision is ported from the revision D10247 based on the report from Intel Parallel Inspector which discovered in both main and sub threads trying 
change the boolean value in the function: forceRedraw().
This revision also changed the name of mutex which will be used not only for onoff array , because I don't want to add a new static variable.

Test Plan: Tested by loading multi-wells and pick sets, it works as usual.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D10248