[osgGeo1.2] OpendTect does not come up anymore because of locking problem.
Summary:
(1) OpendTect does not come up anymore because of locking problem. Caused by
Ding's previous commit of MarkerSet(.cpp): "This revision also changed the
name of mutex which will be used not only for onoff array , because I don't
want to add a new static variable." See Phabricator D10248. Apparently,
sharing this mutex for two diferent things is not a good idea. Adding a
new one instead seems to do the job, and is more analog to the commit in
osgGeo-master, which does not have this problem.

Test Plan:
It compiles and now OpendTect does come up again. No clue how to test the
new lock itself, but I trust that has been tested by Ding.

Reviewers: kristofer

Reviewed By: kristofer

Subscribers: ding

Differential Revision: https://phabricator.dgbes.com:8443/D10361