[osgGeo] Get dragger info to customize plane dragging in OD (OD-262)
Summary:
Dragger info functionality to customize plane dragging in OpendTect.
OSG's dragging behaviour is based on the 3D scene co-ordinates of the
mouse pointer, which is highly non-linear at particular viewing angles.
Where necessary, OpendTect will base the dragging behaviour on the 2D
screen co-ordinates of the mouse pointer instead. In order to provide
reasonable mouse-to-dragger displacements, the lengths of the returned
screen projection vectors are chosen to be proportional to the number
of pixels occupied by the main axes of the plane.

Test Plan:
Tested in OpendTect using left- and right-handed co-ordinate systems,
using perspective and orthonormal projection, and for many different
camera angles. It took a lot of trial-and-error to get everything
consistent.

Reviewers: kristofer

Reviewed By: kristofer

Subscribers: nanne

Differential Revision: https://phabricator.dgbes.com:8443/D10677