[osgGeo1.2] Parallelization of encodeBaseChannelPower(.) (part of OD-660) ABI PRESERVATION: (UN)REFFING THREADGROUPS IN CONSTRUCTORS/DESTRUCTOR
Summary:
Parallelization of encodeBaseChannelPower(.) function for speedup of OD
inl/crl/z/rl dragging with preloaded data. Turned out to be one of the
speed bottlenecks when the images get huge. Moved all your ThreadGroup
functionality to template classes in order to be used both for composite
texture creation and power encoding, and probably more in the future.
Same threads are used by different class instances, but I do not see
an easy way to use the same thread for different tasks as well.

Test Plan:
Tested in OpendTect. Temporarily disabled shading to assure composite texture
creation still works. The power encoding reached 25% speedup for an average
sized inline on a dual-core computer. The overhead of parallellization does
count for such "light-weight" operations. When using the class variables in
the iteration loop, there was no speedup at all in comparison with the
original static encodeBaseChannelPower(.) function. It turns out beneficial
to use only local variables in the iteration loop to have them on the stack.

Reviewers: kristofer

Reviewed By: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D10962