[od]OD-1527 Crashed using lasso tool while horizon picking and auto tracking
Summary:
After investigation, the crash is caused by the operation of camera clone which brings memory problem when
working on big horizon.
actually the copy purpose of polygon selection is to be used in doing multi selections. So for each of selection
we only need to know the selected polygon coords and how to check if a position is inside or not. we don't need doing camera
render. Based on this idea, I made this solution.

Test Plan:
do following steps in the two surveys mentioned in the Jira OD-1527

    resumed tracking of existing horizon
    auto-tracked a large part
    selected points with lasso tool and removed them, then repeated so multiple times, also tested holding Ctrl to select points within several polygons
    auto-tracked more
    used lasso tool and removed selection
    saved horizon as a new one (called it 'saveas')
    removed it from Tree
    added it back in the Tree and resumed tracking
    then again used auto-tracking, and lasso selection and removal

It works fine.

Reviewers: ranojay

Reviewed By: ranojay

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D11465