[od6.2]OD-1527 Crashed using lasso tool while horizon picking and auto tracking
Summary: This is the same as revision D11463

Test Plan:
After investigation, the crash is caused by the operation of camera clone which brings memory problem when
working on big horizon.
actually the copy purpose of polygon selection is to be used in doing multi selections. So for each of selection
we only need to know the selected polygon coords and how to check if a position is inside or not. we don't need doing camera
render. Based on this idea, I made this solution.

Reviewers: ranojay

Reviewed By: ranojay

Subscribers: kristofer

Differential Revision: https://phabricator.dgbes.com:8443/D11464