[osgGeo-trunk] Possible bugfix for ION crash in duplicated OSG code.
Summary:
Possible bugfix for ION crash in duplicated OSG code. Current code was created
before Robert Osfield committed this revision to osgSim/ScalarBar.cpp:
https://github.com/openscenegraph/OpenSceneGraph/commit/dd996a328979614d7c6b5778b2e4f9c8af37ecfd
Only a small chance that it will solve the ION crash, but this code is indeed
called at OpendTect startup, fonts do have something to do with the text
display calls on the call stack, and Robert Osfield mentions it solves
a potential threading bug.

Test Plan: It compiles, and the scalar bar in OpendTect is displayed correctly.

Reviewers: ranojay

Reviewed By: ranojay

Differential Revision: https://phabricator.dgbes.com:8443/D11591